Fix issue when parsing back face data from a sensor where all elements would appear to be 0.
Issue was to do with the OpenCV mat type - we needed a floating point type, but were giving a signed integer type

+++ projects/Core/src/com/gpigc/core/analysis/engine/FaceAnalysisEngine.java
@@ -129,7 +129,7 @@ public class FaceAnalysisEngine extends AnalysisEngine {
 	public static Mat parseFace(String faceMatrixStr) {
 		// Parse face matrix
 		String[] elements = faceMatrixStr.split(",");
-		Mat faceMatrix = new Mat(new Size(elements.length, 1), CvType.CV_32SC1);
+		Mat faceMatrix = new Mat(new Size(elements.length, 1), CvType.CV_64FC1);
 		for (int i = 0; i < elements.length; i++) {
 			double elementValue = Double.parseDouble(elements[i]);
 			System.err.println("Parsing element \""+elements[i]+"\" as "+
